add api for skill service & modify user api

+++ spring-boot-rest-server/src/main/java/com/yyfolium/springbootrestserver/header/HeaderRepository.java
@@ -2,5 +2,5 @@ package com.yyfolium.springbootrestserver.header;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HeaderRepository extends JpaRepository<Header, Integer> {
+public interface HeaderRepository extends JpaRepository<Header, Long> {
 }
