+++ spring-boot-rest-server/src/main/java/com/yyfolium/springbootrestserver/job/JobRepository.java
@@ -2,5 +2,5 @@ package com.yyfolium.springbootrestserver.job;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JobRepository extends JpaRepository<Job, Integer> {
+public interface JobRepository extends JpaRepository<Job, Long> {
 }
