+++ spring-boot-rest-server/src/main/java/com/yyfolium/springbootrestserver/portfolio/PortfolioImageRepository.java
@@ -2,5 +2,5 @@ package com.yyfolium.springbootrestserver.portfolio;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PortfolioImageRepository extends JpaRepository<PortfolioImage, Integer> {
+public interface PortfolioImageRepository extends JpaRepository<PortfolioImage, Long> {
 }
