+++ spring-boot-rest-server/src/main/java/com/yyfolium/springbootrestserver/portfolio/PortfolioMenuRepository.java
@@ -2,5 +2,5 @@ package com.yyfolium.springbootrestserver.portfolio;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PortfolioMenuRepository extends JpaRepository<PortfolioMenu, Integer> {
+public interface PortfolioMenuRepository extends JpaRepository<PortfolioMenu, Long> {
 }
