+++ spring-boot-rest-server/src/main/java/com/yyfolium/springbootrestserver/portfolio/PortfolioProjectRepository.java
@@ -2,5 +2,5 @@ package com.yyfolium.springbootrestserver.portfolio;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PortfolioProjectRepository extends JpaRepository<PortfolioProject, Integer> {
+public interface PortfolioProjectRepository extends JpaRepository<PortfolioProject, Long> {
 }
