+++ spring-boot-rest-server/src/main/java/com/yyfolium/springbootrestserver/resume/ResumeRepository.java
@@ -2,5 +2,5 @@ package com.yyfolium.springbootrestserver.resume;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ResumeRepository extends JpaRepository<Resume, Integer> {
+public interface ResumeRepository extends JpaRepository<Resume, Long> {
 }
